+++ crates/warpgate-api/src/lib.rs
@@ -1,6 +1,7 @@
 mod host;
 mod host_funcs;
 mod virtual_path;
+mod macros;
 
 pub use host::*;
 pub use host_funcs::*;
@@ -10,20 +11,3 @@ api_struct!(
     /// Represents an empty input.
     pub struct EmptyInput {}
 );
-
-#[macro_export]
-macro_rules! api_struct {
-    ($struct:item) => {
-        #[derive(Clone, Debug, Default, serde::Deserialize, Eq, PartialEq, serde::Serialize)]
-        #[serde(default)]
-        $struct
-    };
-}
-
-#[macro_export]
-macro_rules! api_enum {
-    ($struct:item) => {
-        #[derive(Clone, Debug, serde::Deserialize, Eq, PartialEq, serde::Serialize)]
-        $struct
-    };
-}
